requirements.txt
langchain-community==0.0.29
sentence-transformers==2.2.2
pypdf==3.17.0
faiss-cpu==1.7.4
streamlit==1.29.0
python-dotenv==1.0.0
openai==1.3.9
torch==2.1.1
transformers==4.35.2
chromadb==0.4.22
tiktoken==0.5.2
